+++ moonmining/models.py
@@ -73,6 +73,16 @@ class NotificationType(str, Enum):
             cls.MOONMINING_LASER_FIRED,
         }
 
+    @classproperty
+    def with_products(cls) -> set:
+        """Return all notification types with have products."""
+        return {
+            cls.MOONMINING_AUTOMATIC_FRACTURE,
+            cls.MOONMINING_EXTRACTION_FINISHED,
+            cls.MOONMINING_EXTRACTION_STARTED,
+            cls.MOONMINING_LASER_FIRED,
+        }
+
 
 class OreRarityClass(models.IntegerChoices):
     """Rarity class of an ore"""
@@ -322,6 +332,7 @@ class Extraction(models.Model):
         "Refinery", on_delete=models.CASCADE, related_name="extractions"
     )
     started_at = models.DateTimeField(help_text="when this extraction was started")
+    # TODO: Add db_index
     # normal properties
     auto_fracture_at = models.DateTimeField(
         help_text="when this extraction will be automatically fractured",
@@ -461,6 +472,56 @@ class Extraction(models.Model):
         self.value = self.calc_value()
         self.is_jackpot = self.calc_is_jackpot()
         self.save()
+
+    def to_calculated_extraction(self) -> CalculatedExtraction:
+        """Generate a calculated extraction from this extraction."""
+
+        def _products_to_calculated_products():
+            return [
+                CalculatedExtractionProduct(
+                    ore_type_id=obj.ore_type_id, volume=obj.volume
+                )
+                for obj in self.products.all()
+            ]
+
+        params = {"refinery_id": self.refinery_id}
+        if self.status == self.Status.STARTED:
+            params.update(
+                {
+                    "status": CalculatedExtraction.Status.STARTED,
+                    "chunk_arrival_at": self.chunk_arrival_at,
+                    "auto_fracture_at": self.auto_fracture_at,
+                    "started_at": self.started_at,
+                    "started_by": self.started_by,
+                    "products": _products_to_calculated_products(),
+                }
+            )
+        elif self.status == self.Status.READY:
+            params.update(
+                {
+                    "status": CalculatedExtraction.Status.READY,
+                    "auto_fracture_at": self.auto_fracture_at,
+                    "products": _products_to_calculated_products(),
+                }
+            )
+        elif self.status == self.Status.COMPLETED:
+            params.update(
+                {
+                    "fractured_by": self.fractured_by,
+                    "fractured_at": self.fractured_at,
+                    "status": CalculatedExtraction.Status.COMPLETED,
+                    "products": _products_to_calculated_products(),
+                }
+            )
+        elif self.status == self.Status.CANCELED:
+            params.update(
+                {
+                    "status": CalculatedExtraction.Status.CANCELED,
+                    "canceled_at": self.canceled_at,
+                    "canceled_by": self.canceled_by,
+                }
+            )
+        return CalculatedExtraction(**params)
 
 
 class ExtractionProduct(models.Model):
@@ -740,8 +801,19 @@ class Moon(models.Model):
             return True
         return False
 
-    def update_products_from_notification(self, notification):
-        ...
+    def update_products_from_latest_extraction(
+        self, overwrite_survey: bool = False
+    ) -> bool:
+        try:
+            extraction = self.refinery.extractions.order_by("-started_at").first()
+        except ObjectDoesNotExist:
+            return None
+        if not extraction:
+            return None
+        calculated_extraction = extraction.to_calculated_extraction()
+        return self.update_products_from_calculated_extraction(
+            calculated_extraction, overwrite_survey=overwrite_survey
+        )
 
 
 class MoonProduct(models.Model):
@@ -826,30 +898,8 @@ class Notification(models.Model):
             self.notif_type,
         )
 
-    def to_calculated_extraction_status(self) -> CalculatedExtraction.Status:
-        status_map = {
-            NotificationType.MOONMINING_EXTRACTION_STARTED: (
-                CalculatedExtraction.Status.STARTED
-            ),
-            NotificationType.MOONMINING_EXTRACTION_CANCELLED: (
-                CalculatedExtraction.Status.CANCELED
-            ),
-            NotificationType.MOONMINING_EXTRACTION_FINISHED: (
-                CalculatedExtraction.Status.READY
-            ),
-            NotificationType.MOONMINING_LASER_FIRED: (
-                CalculatedExtraction.Status.COMPLETED
-            ),
-            NotificationType.MOONMINING_AUTOMATIC_FRACTURE: (
-                CalculatedExtraction.Status.COMPLETED
-            ),
-        }
-        try:
-            return status_map[self.notif_type]
-        except KeyError:
-            return CalculatedExtraction.Status.UNDEFINED
-
     def to_calculated_extraction(self) -> CalculatedExtraction:
+        """Generate a calculated extraction from this notification."""
         params = {"refinery_id": self.details["structureID"]}
         if self.notif_type == NotificationType.MOONMINING_EXTRACTION_STARTED:
             params.update(
@@ -880,6 +930,8 @@ class Notification(models.Model):
         ):
             params.update(
                 {
+                    "fractured_by": self.details.get("firedBy"),
+                    "fractured_at": self.timestamp,
                     "status": CalculatedExtraction.Status.COMPLETED,
                     "products": CalculatedExtractionProduct.create_list_from_dict(
                         self.details["oreVolumeByType"]
@@ -890,6 +942,7 @@ class Notification(models.Model):
             params.update(
                 {
                     "status": CalculatedExtraction.Status.CANCELED,
+                    "canceled_at": self.timestamp,
                     "canceled_by": self.details.get("cancelledBy"),
                 }
             )
@@ -1184,21 +1237,7 @@ class Owner(models.Model):
                 refinery.update_moon_from_eve_id(notif.details["moonID"])
             for notif in notifications_for_refinery.order_by("timestamp"):
                 if notif.notif_type == NotificationType.MOONMINING_EXTRACTION_STARTED:
-                    extraction = CalculatedExtraction(
-                        refinery_id=refinery.id,
-                        status=CalculatedExtraction.Status.STARTED,
-                        chunk_arrival_at=ldap_time_2_datetime(
-                            notif.details["readyTime"]
-                        ),
-                        auto_fracture_at=ldap_time_2_datetime(
-                            notif.details["autoTime"]
-                        ),
-                        started_at=notif.timestamp,
-                        started_by=notif.details.get("startedBy"),
-                        products=CalculatedExtractionProduct.create_list_from_dict(
-                            notif.details["oreVolumeByType"]
-                        ),
-                    )
+                    extraction = notif.to_calculated_extraction()
                     if refinery.moon.update_products_from_calculated_extraction(
                         extraction
                     ):
@@ -1269,16 +1308,7 @@ class Owner(models.Model):
                         notif.notif_type
                         == NotificationType.MOONMINING_EXTRACTION_FINISHED
                     ):
-                        extraction = CalculatedExtraction(
-                            refinery_id=refinery.id,
-                            status=CalculatedExtraction.Status.READY,
-                            auto_fracture_at=ldap_time_2_datetime(
-                                notif.details["autoTime"]
-                            ),
-                            products=CalculatedExtractionProduct.create_list_from_dict(
-                                notif.details["oreVolumeByType"]
-                            ),
-                        )
+                        extraction = notif.to_calculated_extraction()
 
             if extraction:
                 updated = Extraction.objects.update_from_calculated(extraction)