+++ moonmining/tests/test_models.py
@@ -202,6 +202,31 @@ class TestEveOreTypeProfileUrl(NoSocketsTestCase):
 #         self.assertFalse(result)
 
 
+class TestExtraction(NoSocketsTestCase):
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+        load_eveuniverse()
+        load_allianceauth()
+
+    def test_should_convert_to_calculated_extraction(self):
+        # given
+        refinery = RefineryFactory()
+        my_map = [
+            (Extraction.Status.STARTED, CalculatedExtraction.Status.STARTED),
+            (Extraction.Status.CANCELED, CalculatedExtraction.Status.CANCELED),
+            (Extraction.Status.READY, CalculatedExtraction.Status.READY),
+            (Extraction.Status.COMPLETED, CalculatedExtraction.Status.COMPLETED),
+        ]
+        for in_status, out_status in my_map:
+            with self.subTest(status=in_status):
+                extraction = ExtractionFactory(status=in_status, refinery=refinery)
+                # when
+                obj = extraction.to_calculated_extraction()
+                # then
+                self.assertEqual(obj.status, out_status)
+
+
 class TestMoonUpdateValue(NoSocketsTestCase):
     @classmethod
     def setUpClass(cls):
@@ -376,6 +401,7 @@ class TestMoonOverwriteProducts(NoSocketsTestCase):
         super().setUpClass()
         load_eveuniverse()
         helpers.generate_market_prices()
+        load_allianceauth()
 
     def test_should_overwrite_existing_estimates(self):
         # given
@@ -433,6 +459,17 @@ class TestMoonOverwriteProducts(NoSocketsTestCase):
         self.assertFalse(result)
         self.assertTrue(moon.products.exists())
 
+    def test_should_overwrite_products_from_latest_extraction(self):
+        # given
+        moon = MoonFactory()
+        refinery = RefineryFactory(moon=moon)
+        ExtractionFactory(refinery=refinery)
+        moon.products.all().delete()
+        # when
+        moon.update_products_from_latest_extraction()
+        # then
+        self.assertGreater(moon.products.count(), 0)
+
 
 class TestNotification(NoSocketsTestCase):
     @classmethod
@@ -440,24 +477,6 @@ class TestNotification(NoSocketsTestCase):
         super().setUpClass()
         load_eveuniverse()
         load_allianceauth()
-
-    def test_should_return_status_for_notifications(self):
-        # given
-        refinery = RefineryFactory()
-        my_map = [
-            (Extraction.Status.STARTED, CalculatedExtraction.Status.STARTED),
-            (Extraction.Status.CANCELED, CalculatedExtraction.Status.CANCELED),
-            (Extraction.Status.READY, CalculatedExtraction.Status.READY),
-            (Extraction.Status.COMPLETED, CalculatedExtraction.Status.COMPLETED),
-        ]
-        for in_status, out_status in my_map:
-            with self.subTest(status=in_status):
-                extraction = ExtractionFactory(status=in_status, refinery=refinery)
-                notification = NotificationFactory(extraction=extraction)
-                # when/then
-                self.assertEqual(
-                    notification.to_calculated_extraction_status(), out_status
-                )
 
     def test_should_convert_to_calculated_extraction(self):
         # given